# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/profile:
    get:
      operationId: getStudentProfile
      tags: [student-profile]
      summary: Get own profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateStudentProfile
      tags: [student-profile]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /student/restaurants:
    get:
      operationId: listStudentRestaurants
      tags: [student-restaurants]
      summary: List active restaurants
      security:
        - bearerAuth: []
      parameters:
        - name: campusId
          in: query
          schema:
            type: string
        - name: slot
          in: query
          schema:
            type: string
            enum: [lunch, dinner, both]
      responses:
        "200":
          description: Restaurant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRestaurant"
  /student/restaurants/{restaurantId}:
    get:
      operationId: getStudentRestaurant
      tags: [student-restaurants]
      summary: Get restaurant detail with packages and meals
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restaurant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRestaurantDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /student/restaurants/{restaurantId}/rate:
    post:
      operationId: rateRestaurant
      tags: [student-restaurants]
      summary: Submit a rating for a restaurant
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateRestaurantBody"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /student/orders:
    get:
      operationId: listStudentOrders
      tags: [student-orders]
      summary: List meal orders
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentOrder"
  /student/orders/{orderId}/cancel:
    post:
      operationId: cancelStudentOrder
      tags: [student-orders]
      summary: Cancel a single meal order
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancellation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrderResponse"
        "400":
          description: Cannot cancel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    StudentProfile:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
        campusId:
          type: string
          nullable: true
        campusName:
          type: string
          nullable: true
        foodPreference:
          type: string
          enum: [veg, non-veg, egg, jain]
        address:
          type: string
          nullable: true
        walletBalance:
          type: number
        isProfileComplete:
          type: boolean
        createdAt:
          type: string
      required: [id, phone, name, foodPreference, walletBalance, isProfileComplete]
...
    StudentUpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        campusId:
          type: string
        campusName:
          type: string
        foodPreference:
          type: string
          enum: [veg, non-veg, egg, jain]
        address:
          type: string
...
    StudentRestaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        cuisineType:
          type: string
          nullable: true
        campusId:
          type: string
          nullable: true
        isVeg:
          type: boolean
        lunchAvailable:
          type: boolean
        dinnerAvailable:
          type: boolean
        operatingDays:
          type: array
          items:
            type: string
        maxCapacity:
          type: integer
          nullable: true
        accentColor:
          type: string
          nullable: true
        lunchStartPrice:
          type: number
          nullable: true
        dinnerStartPrice:
          type: number
          nullable: true
        cancelCutoffHour:
          type: integer
        deliveryTime:
          type: string
          nullable: true
        distanceLabel:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
      required: [id, name, isVeg, lunchAvailable, dinnerAvailable, operatingDays, cancelCutoffHour, reviewCount]
    StudentPackage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        durationDays:
          type: integer
        mealSlot:
          type: string
          enum: [lunch, dinner, both]
        pricePerDay:
          type: number
        totalPrice:
          type: number
        discountPercent:
          type: number
        mealsPerDay:
          type: integer
      required: [id, name, durationDays, mealSlot, pricePerDay, totalPrice, discountPercent, mealsPerDay]
    StudentMeal:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        shortDescription:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        vegType:
          type: string
          enum: [veg, non-veg, egg]
        spiceLevel:
          type: string
          enum: [mild, medium, hot]
          nullable: true
        calories:
          type: integer
          nullable: true
        storagePath:
          type: string
          nullable: true
        lunchAvailable:
          type: boolean
        dinnerAvailable:
          type: boolean
      required: [id, name, vegType, lunchAvailable, dinnerAvailable]
    StudentRestaurantDetail:
      allOf:
        - $ref: "#/components/schemas/StudentRestaurant"
        - type: object
          properties:
            packages:
              type: array
              items:
                $ref: "#/components/schemas/StudentPackage"
            meals:
              type: array
              items:
                $ref: "#/components/schemas/StudentMeal"
            ratingsBreakdown:
              type: object
              nullable: true
              properties:
                totalReviews:
                  type: integer
                overall:
                  type: number
                  nullable: true
                foodQuality:
                  type: number
                  nullable: true
                packaging:
                  type: number
                  nullable: true
                delivery:
                  type: number
                  nullable: true
                valueForMoney:
                  type: number
                  nullable: true
                hygiene:
                  type: number
                  nullable: true
          required: [packages, meals]
    RateRestaurantBody:
      type: object
      properties:
        foodQuality:
          type: number
          minimum: 1
          maximum: 5
        packaging:
          type: number
          minimum: 1
          maximum: 5
        delivery:
          type: number
          minimum: 1
          maximum: 5
        valueForMoney:
          type: number
          minimum: 1
          maximum: 5
        hygiene:
          type: number
          minimum: 1
          maximum: 5
        communication:
          type: number
          minimum: 1
          maximum: 5
        overall:
          type: number
          minimum: 1
          maximum: 5
        note:
          type: string
          nullable: true
      required: [foodQuality, packaging, delivery, valueForMoney, hygiene, overall]
...
    StudentOrder:
      type: object
      properties:
        id:
          type: string
        subscriptionId:
          type: string
        restaurantId:
          type: string
        packageName:
          type: string
        mealSlot:
          type: string
          enum: [lunch, dinner]
        scheduledDate:
          type: string
        status:
          type: string
          enum: [scheduled, locked, accepted, preparing, ready, delivered, cancelled, no_show]
        pricePerDay:
          type: number
        freeCancelUntil:
          type: string
        isLocked:
          type: boolean
        cancelStatus:
          type: string
          enum: [free, late, full, none]
      required: [id, subscriptionId, restaurantId, packageName, mealSlot, scheduledDate, status, pricePerDay, freeCancelUntil, isLocked, cancelStatus]
    CancelOrderResponse:
      type: object
      properties:
        type:
          type: string
          enum: [free, late, full]
        fee:
          type: number
        refund:
          type: number
        message:
          type: string
      required: [type, fee, refund, message]
...